Initial version of stream with read subtest
Stream is great but it gets confusing to users when we try to
explain why triad score is so far below theoretical bandwidth.
The write of the output arrays isn't included in the bw.
This is okay if you are doing non-temporal stores but we usually
don't do non-temporal stores.
Add a pure-read subtest which gets about the same bw as Intel mlc
utility or my own spin.x pure-read bw.